+++ lib/flutter_helper_kit.dart
@@ -67,6 +67,7 @@ export './widgets/timer_builder.dart';
 export './widgets/separated_column.dart';
 export './widgets/space/sliver_space.dart';
 export './widgets/space/space.dart';
+export './widgets/flutter_list_view.dart';
 
 ///CustomPainter
 export 'custom_painter/google_logo_painter.dart';