+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+## 0.0.7
+* TODO: add FlutterListView widget and Change ReadMe file..
 
+## 0.0.6
+* TODO: add FlutterListView widget.
 
 ## 0.0.5
 * TODO: Update Readme Example Images Path.